This version includes the first attempt to change the way units are shown for numeric data types. It's not complete, but now, we can use arbitrary exponents instead of the metric prefixes only.
The duplication of marks via user gestures is implemented: Mouse down and hold + drag gesture to the intended location of the copy + mouse up. This change introduced a state for the marks called copyingMode. In addition, when a mark is being duplicated by the user though interaction, the created copy is stored in the original mark object as a variable that will be accessed from the different methods that are needed to complete the touch interaction. The name of this variable is currentCopy