A bug in Mark cloning has been fixed: When a mark was copied, the copy did not locate its elements (visual properties, background rect) as it was moved. Because of this, incoming connectors where mislocated, as their ending location was always set to the original mark.
The code to download images taken from web pages has is now working with paths that are automatically defined. There is no need to define them by hard.

A new input data icon is available on the main menu bar of iVoLVR. It will allow to manipulate elements of SVG files.