Homogeneity can now be determined for arrays of strings that should be converted to collections of visual values. There is a function that individually guesses the best data representation for the given string. Based on this function, another one decides which type should be applied to the entire collection. This is already working when selecting rows from tables of HTML documents.
The Data widget has been modified to show, using the coloring system of the iVoLVR data types, the type of the collections of data it holds. For instance, if one column in the file contain numbers, then, the data widget shows a data variable colored with the colors of the numeric visual data.

This commit also includes several chnages, that did not work at the end, to cancel the "going back/forward" of the touch events on the iframe that shows the web pages.