The constructor of the Operator class is modified in such a way that when a saved Operator is loaded from an XML iVoLVR project file, it is rendered with the color of the object that it holds (blue if it has a numeric value, for instance).
Connectors are now saved! This is done through the xmlID attribute for all the fabric objects that are part of a project.

The status whether the canvas has the connectors hidden or not is also saved to the project file so that the corresponding value is recovered when the project is open. The value of the checkbox of the tool bar that controls the connectors visibility is also adjusted properly.

Locators are now saved in their very first version. A decision should be made on how to store the values of the location visual properties and how they should interact with locators.