Locators and Number Generators are now saved. There is an issue with synchronization, because locators always refer to marks, but some of them (specially SVG marks) might not be on the canvas already when the locator tries to do the linking. This can also happen with connectors, as they also refer to objects that might not be on the canvas.
Another issue is present when a connection involves a non visible object. For instance, when the the destination object is any of the position visual properties of a mark that belongs to a locator. In this case, the search function does not return the proper object, as this function searches on the canvas objects. A decision on this issue should be made because, otherwise, the connection is not drawn.