Mappers can now be saved and loaded.
Color extractors can now give their shape to be used in marks. The values held by their visual properties are now computed during its creation.

Fixed a bug that did not allow to change the shape of marks because of an non-undefined checking.

Now, when a collection is input to a mapper, the input and output points go to the upper part of the element so that there is no confusion on the value being evaluated.