Color samplers are now improved. The errors originated for very short traced paths are now handled properly in such a way that, if the removal of self intersections fails, the original offset polygon is used and no problems are propagated.
There is a major improvement in the efficiency of the elements that need to be brought to the front of the canvas. A bringToFront function was implemented to avoid the one provided by the canvas object, as the latter one refreshed the canvas, introducing delays. This is now noticeable when having color samplers with a large number of sampling points. The same function has now being used in the entire system to avoid undesired canvas refreshing calls.