The interfaz has been fully modified to have a "Extractors" section in the right menu. The "mutual exclusion" on the functionalities of the different modes is now implemented. There are specific rules for the double tap gesture that activates the floodfill algorithm, as it is not mutually exclusive with the panning of the canvas or the disconnecting mode.
Any menu open at the moment of any interaction with the canvas is now closed with the use of tow functions: showMenu and hideMenu.

The corresponding behaviour for each mode is now working as it should be. Some testing has been done, though some more additional tests would be recommended.